LayoutTests/imported/w3c/web-platform-tests/html/interaction/focus/the-autofocus-attribute/same-origin-autofocus-expected.txt

CONSOLE MESSAGE: Blocked autofocusing on a form control in a cross-origin subframe.


PASS Autofocus should not work in the same origin grand child iframe
